requirements.txt
fastapi~=0.88.0
SQLAlchemy~=1.4.45
passlib~=1.7.4
python-dotenv~=0.21.0
uvicorn~=0.20.0
pydantic~=1.10.2
shortuuid~=1.0.11
jose~=1.0.0
starlette~=0.22.0
python-decouple~=3.4
mysql-connector-python~=8.0.27
FastAPI-SQLAlchemy~=0.0.3
